+++ lib/verb.js
@@ -1,0 +1,11 @@
+'use strict'
+
+var Verb = function(){
+this.awake = true;
+this.beaten = true;
+this.hidden = true;
+this.meant = true;
+this.said = true;
+};
+
+module.exports = Verb;
